refactor:test
- 整合部分测试代码
