+++ morn-boot-test/src/test/java/site/morn/test/TestUser.java
@@ -1,10 +1,13 @@
-package site.morn.boot.jpa;
+package site.morn.test;
 
+import java.io.Serializable;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
+import javax.validation.constraints.NotNull;
 import lombok.Data;
+import site.morn.validate.group.Add;
 
 /**
  * 测试用户
@@ -14,15 +17,18 @@ import lombok.Data;
  */
 @Data
 @Entity
-public class TestUser {
+public class TestUser implements Serializable {
+
 
   @Id
   @GeneratedValue
   private Long id;
 
   @Column
+  @NotNull
   private String username;
 
   @Column
+  @NotNull(groups = Add.class)
   private String password;
 }
