+++ morn-boot-test/src/test/java/site/morn/validate/TestValidateController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 import site.morn.rest.RestBuilders;
 import site.morn.rest.RestMessage;
 import site.morn.rest.RestModel;
+import site.morn.test.TestUser;
 import site.morn.validate.group.Add;
 import site.morn.validate.group.Update;
